Move all the mocks to their own module
