+++ src/pd/client.rs
@@ -163,7 +163,7 @@ impl PdRpcClient<TikvConnect, Cluster> {
 }
 
 impl<KvC: KvConnect + Send + Sync + 'static, Cl> PdRpcClient<KvC, Cl> {
-    fn new<MakeKvC, MakePd>(
+    pub fn new<MakeKvC, MakePd>(
         config: &Config,
         kv_connect: MakeKvC,
         pd: MakePd,
@@ -216,131 +216,9 @@ impl<KvC: KvConnect + Send + Sync + 'static, Cl> PdRpcClient<KvC, Cl> {
 #[cfg(test)]
 pub mod test {
     use super::*;
-    use crate::kv_client::{KvClient, KvRequest};
-    use crate::Error;
+    use crate::mock::*;
 
     use futures::executor;
-    use futures::future::{ready, BoxFuture};
-    use grpcio::CallOption;
-    use kvproto::metapb;
-
-    // FIXME move all the mocks to their own module
-    #[derive(Clone, Eq, PartialEq, Debug)]
-    pub struct MockKvClient {
-        addr: String,
-    }
-
-    impl KvClient for MockKvClient {
-        fn dispatch<T: KvRequest>(
-            &self,
-            _request: &T::RpcRequest,
-            _opt: CallOption,
-        ) -> BoxFuture<'static, Result<T::RpcResponse>> {
-            unreachable!()
-        }
-    }
-
-    pub struct MockKvConnect;
-
-    impl KvConnect for MockKvConnect {
-        type KvClient = MockKvClient;
-
-        fn connect(&self, address: &str) -> Result<Self::KvClient> {
-            Ok(MockKvClient {
-                addr: address.to_owned(),
-            })
-        }
-    }
-
-    pub struct MockCluster;
-
-    pub struct MockPdClient;
-
-    impl PdClient for MockPdClient {
-        type KvClient = MockKvClient;
-
-        fn map_region_to_store(
-            self: Arc<Self>,
-            region: Region,
-        ) -> BoxFuture<'static, Result<Store<Self::KvClient>>> {
-            Box::pin(ready(Ok(Store::new(
-                region,
-                MockKvClient {
-                    addr: String::new(),
-                },
-                Duration::from_secs(60),
-            ))))
-        }
-
-        fn region_for_key(&self, key: &Key) -> BoxFuture<'static, Result<Region>> {
-            let bytes: &[_] = key.into();
-            let region = if bytes.is_empty() || bytes[0] < 10 {
-                Self::region1()
-            } else {
-                Self::region2()
-            };
-
-            Box::pin(ready(Ok(region)))
-        }
-        fn region_for_id(&self, id: RegionId) -> BoxFuture<'static, Result<Region>> {
-            let result = match id {
-                1 => Ok(Self::region1()),
-                2 => Ok(Self::region2()),
-                _ => Err(Error::region_not_found(id, None)),
-            };
-
-            Box::pin(ready(result))
-        }
-
-        fn get_timestamp(self: Arc<Self>) -> BoxFuture<'static, Result<Timestamp>> {
-            unimplemented!()
-        }
-    }
-
-    impl MockPdClient {
-        fn region1() -> Region {
-            let mut region = Region::default();
-            region.region.id = 1;
-            region.region.set_start_key(vec![0]);
-            region.region.set_end_key(vec![10]);
-
-            let mut leader = metapb::Peer::default();
-            leader.store_id = 41;
-            region.leader = Some(leader);
-
-            region
-        }
-
-        fn region2() -> Region {
-            let mut region = Region::default();
-            region.region.id = 2;
-            region.region.set_start_key(vec![10]);
-            region.region.set_end_key(vec![250, 250]);
-
-            let mut leader = metapb::Peer::default();
-            leader.store_id = 42;
-            region.leader = Some(leader);
-
-            region
-        }
-    }
-
-    fn pd_rpc_client() -> PdRpcClient<MockKvConnect, MockCluster> {
-        let config = Config::default();
-        PdRpcClient::new(
-            &config,
-            |_, _| MockKvConnect,
-            |e, sm| {
-                Ok(RetryClient::new_with_cluster(
-                    e,
-                    sm,
-                    config.timeout,
-                    MockCluster,
-                ))
-            },
-        )
-        .unwrap()
-    }
 
     #[test]
     fn test_kv_client() {
